requirements.txt
asgiref==3.8.1
Django==5.2.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
djangorestframework==3.15.0
djangorestframework-simplejwt==4.3.0
